refactor: update client and config files
This commit updates the client.ts and config.ts files in the .vuepress directory.

In client.ts, the import statement for the "vuepress-theme-hope/presets/shinning-feature-panel.scss" module is added. Additionally, the setupRunningTimeFooter function from the "vuepress-theme-hope/presets/footerRunningTime.js" module is called with specific configurations.

In config.ts, aliases are added for two components: "@theme-hope/modules/blog/components/BlogHero" and "@theme-hope/components/HomePage". These aliases resolve to their respective paths in the project.

The changes aim to improve the functionality and styling of the VuePress blog.