feat: Add essay about the philosophy of life
Added a new essay titled "命里有时终须有" to the blog/essays directory. This essay explores the philosophy of life and reflects on the idea that some things are destined to happen while others are not. It includes an image named "966519d02c8ca050dc064d479fb711e.jpg" and discusses the author's experience of winning a prize in a game. The essay emphasizes the importance of accepting the natural course of events and the unexpected rewards that can come from it.